Add files via upload
This repo is the solution for a contest NebulaNet, that required us to perform a basic classification task on a dataset containing some information on stars that was to be used to predict their type. train.csv is the training dataset, and test.csv is the test dataset. samplesolutions.csv is the format in which the contest required solutions, and Solution.csv is the prediction of our code exported to a CSV file. Solution.ipynb contains the code written for the purpose.